restructured js
