+++ src/components/shared/ChatRoom.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import * as Constants from './common/constants';
+import * as Constants from './../../common/constants';
 import { connect } from 'react-redux';
