+++ src/redux/reducers/setRoom.js
