+++ src/redux/reducers/setRoomState.js
