+++ src/redux/thunks/runApplyResourceOnVirus.js
