+++ src/redux/thunks/runRemoveMobileUrlBar.js
