+++ src/redux/thunks/runStartGame.js
